auto print cash invoice and cash cancellation invoice after generation
