+++ webroot/js/self-service.js
@@ -272,15 +272,6 @@ foodcoopshop.SelfService = {
 
     renderTimer : function() {
         $('.auto-logout-timer').html(this.currentLogoutTimer);
-    },
-
-    printInvoice : function(file) {
-        var iframe = document.createElement('iframe');
-        iframe.style.visibility = 'hidden';
-        iframe.src = file;
-        document.body.appendChild(iframe);
-        iframe.contentWindow.focus();
-        iframe.contentWindow.print();
     }
 
 };
